+++ CHANGELOG.md
@@ -8,7 +8,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ## [0.5.1] - 2025-03-24
 ### Changed
-- Max length of text fields
+- Max length of text fields in webview
 ### Fixed
 - Error of misstracked command